// src/openapi/model/api-contact.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiContact
 */
export interface ApiContact {
    /**
     *
     * @type {string}
     * @memberof ApiContact
     */
    contactType: string;
    /**
     *
     * @type {string}
     * @memberof ApiContact
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ApiContact
     */
    email: string;
    /**
     *
     * @type {string}
     * @memberof ApiContact
     */
    body: string;
    /**
     *
     * @type {string}
     * @memberof ApiContact
     */
    tel?: string;
}